lyrics/requirements.txt
# Core web framework
fastapi==0.111.0
starlette==0.37.2
uvicorn==0.23.2  # Updated from 0.18.2 for Python 3.13 compatibility

# Templates and file handling
Jinja2==3.1.2
python-multipart==0.0.9

# API and OpenAI integration
openai==1.7.0  # Updated from 0.28.0 (very old version)
requests==2.31.0  # Updated from 2.28.1

# Audio processing
pydub==0.25.1
SpeechRecognition==3.10.0  # Updated from 3.8.1

# Image processing - using a version that's more likely to work with Python 3.13
Pillow==11.2.1 # Downgraded from 10.3.0

# Environment variables
python-dotenv==0.21.0

# Utilities
pydantic==2.7.3
typing_extensions==4.12.2

# Optional dependencies - install if needed
# ujson==5.10.0
# aiohttp==3.9.5

# Commenting out problematic package - use json instead
# orjson==3.10.3

aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi-cli==0.0.4
frozenlist==1.4.1
h11==0.12.0
httpcore==0.15.0
httptools==0.6.1
httpx==0.23.0
idna==3.7
itsdangerous==2.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
pydantic_core==2.18.4
Pygments==2.18.0
PyYAML==6.0.1
rfc3986==1.5.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
tqdm==4.66.4
typer==0.12.3
watchdog==2.2.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4